bok/service_kub/load/requirements.txt
numpy
opencv-python-headless
pymongo
python-dotenv
gridfs